0.0.26
### Added
- on Database.Character.Create fixed inventory.
- OnLogout dispatch despawn player and clear all knowPlayers.
- SMSG_UPDATE_OBJECT.CreateOutOfRangeUpdate
- PlayerManager.DespawnPlayer
- PlayerManager now check range to spawn and despawn Players.

### Changed
- Items Table default values
- Some Summary of files.
- OnItemQueryResponse now check two Ids.
- OnPlayerLogin.inventory now is RealmServerSession.SendInventory
- Fixed some spaces.
- SMSG_ITEM_QUERY_SINGLE_RESPONSE 90% done.

+++ RealmServer/PacketServer/SMSG_PLAYED_TIME.cs
@@ -7,8 +7,8 @@ namespace RealmServer.PacketServer
     {
         public SMSG_PLAYED_TIME(Characters character) : base(RealmEnums.CMSG_PLAYED_TIME)
         {
-            Write((int) 1); // GetTotalPlayedTime
-            Write((int) 1); // GetLevelPlayedTime
+            Write(1); // GetTotalPlayedTime
+            Write(1); // GetLevelPlayedTime
         }
     }
 }
